+++ aclogview/Form1.Designer.cs
@@ -35,6 +35,8 @@ namespace aclogview {
             this.columnHeader4 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
             this.columnHeader6 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
             this.HexOpCode = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
+            this.listviewContextMenu = new System.Windows.Forms.ContextMenuStrip(this.components);
+            this.copyTimeMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.columnHeader13 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
             this.columnHeader15 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
             this.columnHeader14 = ((System.Windows.Forms.ColumnHeader)(new System.Windows.Forms.ColumnHeader()));
@@ -62,6 +64,7 @@ namespace aclogview {
             this.CollapseAll = new System.Windows.Forms.ToolStripMenuItem();
             this.toolStripSeparator1 = new System.Windows.Forms.ToolStripSeparator();
             this.CopyAll = new System.Windows.Forms.ToolStripMenuItem();
+            this.TeleLoc = new System.Windows.Forms.ToolStripMenuItem();
             this.FindID = new System.Windows.Forms.ToolStripMenuItem();
             this.mainMenu = new System.Windows.Forms.MainMenu(this.components);
             this.menuItem_File = new System.Windows.Forms.MenuItem();
@@ -111,6 +114,7 @@ namespace aclogview {
             this.splitContainer_Top.Panel1.SuspendLayout();
             this.splitContainer_Top.Panel2.SuspendLayout();
             this.splitContainer_Top.SuspendLayout();
+            this.listviewContextMenu.SuspendLayout();
             this.objectsContextMenu.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.splitContainer_Bottom)).BeginInit();
             this.splitContainer_Bottom.Panel1.SuspendLayout();
@@ -143,8 +147,8 @@ namespace aclogview {
             // 
             this.splitContainer_Main.Panel2.Controls.Add(this.splitContainer_Bottom);
             this.splitContainer_Main.Panel2.RightToLeft = System.Windows.Forms.RightToLeft.No;
-            this.splitContainer_Main.Size = new System.Drawing.Size(1520, 428);
-            this.splitContainer_Main.SplitterDistance = 25;
+            this.splitContainer_Main.Size = new System.Drawing.Size(1520, 785);
+            this.splitContainer_Main.SplitterDistance = 367;
             this.splitContainer_Main.TabIndex = 0;
             // 
             // splitContainer_Top
@@ -162,7 +166,7 @@ namespace aclogview {
             // 
             this.splitContainer_Top.Panel2.Controls.Add(this.listView_CreatedObjects);
             this.splitContainer_Top.Panel2Collapsed = true;
-            this.splitContainer_Top.Size = new System.Drawing.Size(1520, 25);
+            this.splitContainer_Top.Size = new System.Drawing.Size(1520, 367);
             this.splitContainer_Top.SplitterDistance = 931;
             this.splitContainer_Top.TabIndex = 1;
             // 
@@ -180,13 +184,14 @@ namespace aclogview {
             this.columnHeader13,
             this.columnHeader15,
             this.columnHeader14});
+            this.listView_Packets.ContextMenuStrip = this.listviewContextMenu;
             this.listView_Packets.Dock = System.Windows.Forms.DockStyle.Fill;
             this.listView_Packets.FullRowSelect = true;
             this.listView_Packets.HideSelection = false;
             this.listView_Packets.Location = new System.Drawing.Point(0, 0);
             this.listView_Packets.MultiSelect = false;
             this.listView_Packets.Name = "listView_Packets";
-            this.listView_Packets.Size = new System.Drawing.Size(1516, 21);
+            this.listView_Packets.Size = new System.Drawing.Size(1516, 363);
             this.listView_Packets.TabIndex = 0;
             this.listView_Packets.UseCompatibleStateImageBehavior = false;
             this.listView_Packets.View = System.Windows.Forms.View.Details;
@@ -233,6 +238,21 @@ namespace aclogview {
             // 
             this.HexOpCode.Text = "OpCode";
             this.HexOpCode.Width = 66;
+            // 
+            // listviewContextMenu
+            // 
+            this.listviewContextMenu.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.copyTimeMenuItem});
+            this.listviewContextMenu.Name = "listviewContextMenu";
+            this.listviewContextMenu.Size = new System.Drawing.Size(181, 48);
+            this.listviewContextMenu.Opening += new System.ComponentModel.CancelEventHandler(this.listviewContextMenu_Opening);
+            this.listviewContextMenu.ItemClicked += new System.Windows.Forms.ToolStripItemClickedEventHandler(this.listviewContextMenu_ItemClicked);
+            // 
+            // copyTimeMenuItem
+            // 
+            this.copyTimeMenuItem.Name = "copyTimeMenuItem";
+            this.copyTimeMenuItem.Size = new System.Drawing.Size(180, 22);
+            this.copyTimeMenuItem.Text = "Copy time field";
             // 
             // columnHeader13
             // 
@@ -330,7 +350,7 @@ namespace aclogview {
             // splitContainer_Bottom.Panel2
             // 
             this.splitContainer_Bottom.Panel2.Controls.Add(this.treeView_ParsedData);
-            this.splitContainer_Bottom.Size = new System.Drawing.Size(1520, 399);
+            this.splitContainer_Bottom.Size = new System.Drawing.Size(1520, 414);
             this.splitContainer_Bottom.SplitterDistance = 1074;
             this.splitContainer_Bottom.TabIndex = 0;
             // 
@@ -342,7 +362,7 @@ namespace aclogview {
             this.tabControl1.Location = new System.Drawing.Point(0, 0);
             this.tabControl1.Name = "tabControl1";
             this.tabControl1.SelectedIndex = 0;
-            this.tabControl1.Size = new System.Drawing.Size(1070, 395);
+            this.tabControl1.Size = new System.Drawing.Size(1070, 410);
             this.tabControl1.TabIndex = 0;
             this.tabControl1.Selected += new System.Windows.Forms.TabControlEventHandler(this.tabControl1_Selected);
             // 
@@ -352,7 +372,7 @@ namespace aclogview {
             this.tabHexView.Location = new System.Drawing.Point(4, 22);
             this.tabHexView.Name = "tabHexView";
             this.tabHexView.Padding = new System.Windows.Forms.Padding(3);
-            this.tabHexView.Size = new System.Drawing.Size(1062, 369);
+            this.tabHexView.Size = new System.Drawing.Size(1062, 384);
             this.tabHexView.TabIndex = 0;
             this.tabHexView.Text = "Hex View";
             this.tabHexView.UseVisualStyleBackColor = true;
@@ -377,7 +397,7 @@ namespace aclogview {
             this.hexBox1.ReadOnly = true;
             this.hexBox1.SelectionBackColor = System.Drawing.SystemColors.Highlight;
             this.hexBox1.ShadowSelectionColor = System.Drawing.Color.FromArgb(((int)(((byte)(100)))), ((int)(((byte)(60)))), ((int)(((byte)(188)))), ((int)(((byte)(255)))));
-            this.hexBox1.Size = new System.Drawing.Size(1056, 363);
+            this.hexBox1.Size = new System.Drawing.Size(1056, 378);
             this.hexBox1.StringViewVisible = true;
             this.hexBox1.TabIndex = 1;
             this.hexBox1.UseFixedBytesPerLine = true;
@@ -413,7 +433,7 @@ namespace aclogview {
             this.tabProtocolDocs.Controls.Add(this.protocolWebBrowser);
             this.tabProtocolDocs.Location = new System.Drawing.Point(4, 22);
             this.tabProtocolDocs.Name = "tabProtocolDocs";
-            this.tabProtocolDocs.Size = new System.Drawing.Size(1059, 365);
+            this.tabProtocolDocs.Size = new System.Drawing.Size(1059, 380);
             this.tabProtocolDocs.TabIndex = 1;
             this.tabProtocolDocs.Text = "Protocol Documentation";
             this.tabProtocolDocs.UseVisualStyleBackColor = true;
@@ -424,7 +444,7 @@ namespace aclogview {
             this.protocolWebBrowser.Location = new System.Drawing.Point(0, 0);
             this.protocolWebBrowser.MinimumSize = new System.Drawing.Size(20, 20);
             this.protocolWebBrowser.Name = "protocolWebBrowser";
-            this.protocolWebBrowser.Size = new System.Drawing.Size(1059, 365);
+            this.protocolWebBrowser.Size = new System.Drawing.Size(1059, 380);
             this.protocolWebBrowser.TabIndex = 0;
             // 
             // treeView_ParsedData
@@ -433,7 +453,7 @@ namespace aclogview {
             this.treeView_ParsedData.Dock = System.Windows.Forms.DockStyle.Fill;
             this.treeView_ParsedData.Location = new System.Drawing.Point(0, 0);
             this.treeView_ParsedData.Name = "treeView_ParsedData";
-            this.treeView_ParsedData.Size = new System.Drawing.Size(438, 395);
+            this.treeView_ParsedData.Size = new System.Drawing.Size(438, 410);
             this.treeView_ParsedData.TabIndex = 0;
             this.treeView_ParsedData.AfterSelect += new System.Windows.Forms.TreeViewEventHandler(this.treeView_ParsedData_AfterSelect);
             this.treeView_ParsedData.NodeMouseClick += new System.Windows.Forms.TreeNodeMouseClickEventHandler(this.treeView_ParsedData_NodeMouseClick);
@@ -445,9 +465,10 @@ namespace aclogview {
             this.CollapseAll,
             this.toolStripSeparator1,
             this.CopyAll,
+            this.TeleLoc,
             this.FindID});
             this.parsedContextMenu.Name = "parsedContextMenu";
-            this.parsedContextMenu.Size = new System.Drawing.Size(184, 98);
+            this.parsedContextMenu.Size = new System.Drawing.Size(184, 120);
             this.parsedContextMenu.Opening += new System.ComponentModel.CancelEventHandler(this.parsedContextMenu_Opening);
             this.parsedContextMenu.ItemClicked += new System.Windows.Forms.ToolStripItemClickedEventHandler(this.parsedContextMenu_ItemClicked);
             // 
@@ -474,6 +495,13 @@ namespace aclogview {
             this.CopyAll.ShowShortcutKeys = false;
             this.CopyAll.Size = new System.Drawing.Size(183, 22);
             this.CopyAll.Text = "&Copy All";
+            // 
+            // TeleLoc
+            // 
+            this.TeleLoc.Name = "TeleLoc";
+            this.TeleLoc.Size = new System.Drawing.Size(183, 22);
+            this.TeleLoc.Text = "Copy ACE @teleloc";
+            this.TeleLoc.Visible = false;
             // 
             // FindID
             // 
@@ -681,7 +709,7 @@ namespace aclogview {
             // 
             this.statusStrip.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.toolStripStatus});
-            this.statusStrip.Location = new System.Drawing.Point(0, 452);
+            this.statusStrip.Location = new System.Drawing.Point(0, 809);
             this.statusStrip.Name = "statusStrip";
             this.statusStrip.Size = new System.Drawing.Size(1520, 22);
             this.statusStrip.TabIndex = 1;
@@ -816,7 +844,7 @@ namespace aclogview {
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(1520, 474);
+            this.ClientSize = new System.Drawing.Size(1520, 831);
             this.Controls.Add(this.label1);
             this.Controls.Add(this.HighlightMode_comboBox);
             this.Controls.Add(this.checkBox_ShowObjects);
@@ -846,6 +874,7 @@ namespace aclogview {
             this.splitContainer_Top.Panel2.ResumeLayout(false);
             ((System.ComponentModel.ISupportInitialize)(this.splitContainer_Top)).EndInit();
             this.splitContainer_Top.ResumeLayout(false);
+            this.listviewContextMenu.ResumeLayout(false);
             this.objectsContextMenu.ResumeLayout(false);
             this.splitContainer_Bottom.Panel1.ResumeLayout(false);
             this.splitContainer_Bottom.Panel2.ResumeLayout(false);
@@ -942,6 +971,9 @@ namespace aclogview {
         private System.Windows.Forms.MenuItem menuItem_CheckUpdates;
         private System.Windows.Forms.MenuItem menuItem_Options;
         private System.Windows.Forms.MenuItem menuItem4;
+        private System.Windows.Forms.ToolStripMenuItem TeleLoc;
+        private System.Windows.Forms.ContextMenuStrip listviewContextMenu;
+        private System.Windows.Forms.ToolStripMenuItem copyTimeMenuItem;
         private System.Windows.Forms.ColumnHeader columnHeader13;
         private System.Windows.Forms.ColumnHeader columnHeader15;
         private System.Windows.Forms.ColumnHeader columnHeader14;