+++ aclogview/Tools/OptionsForm.Designer.cs
@@ -1,4 +1,4 @@
-namespace aclogview
+﻿namespace aclogview
 {
     partial class OptionsForm
     {
@@ -37,6 +37,10 @@ namespace aclogview
             this.button_Cancel = new System.Windows.Forms.Button();
             this.toolTip_ProtocolDoc = new System.Windows.Forms.ToolTip(this.components);
             this.groupBox2 = new System.Windows.Forms.GroupBox();
+            this.groupBox4 = new System.Windows.Forms.GroupBox();
+            this.label3 = new System.Windows.Forms.Label();
+            this.label4 = new System.Windows.Forms.Label();
+            this.copyAllPadding = new System.Windows.Forms.TextBox();
             this.displayNodeTooltips = new System.Windows.Forms.CheckBox();
             this.groupBox3 = new System.Windows.Forms.GroupBox();
             this.cbACEStyleHeaders = new System.Windows.Forms.CheckBox();
@@ -44,6 +48,7 @@ namespace aclogview
             this.comboBox_TimeFormat = new System.Windows.Forms.ComboBox();
             this.groupBox1.SuspendLayout();
             this.groupBox2.SuspendLayout();
+            this.groupBox4.SuspendLayout();
             this.groupBox3.SuspendLayout();
             this.SuspendLayout();
             // 
@@ -90,9 +95,9 @@ namespace aclogview
             // 
             // button_OK
             // 
-            this.button_OK.Location = new System.Drawing.Point(252, 233);
+            this.button_OK.Location = new System.Drawing.Point(252, 278);
             this.button_OK.Name = "button_OK";
-            this.button_OK.Size = new System.Drawing.Size(75, 23);
+            this.button_OK.Size = new System.Drawing.Size(75, 24);
             this.button_OK.TabIndex = 1;
             this.button_OK.Text = "OK";
             this.button_OK.UseVisualStyleBackColor = true;
@@ -101,9 +106,9 @@ namespace aclogview
             // button_Cancel
             // 
             this.button_Cancel.DialogResult = System.Windows.Forms.DialogResult.Cancel;
-            this.button_Cancel.Location = new System.Drawing.Point(333, 233);
+            this.button_Cancel.Location = new System.Drawing.Point(333, 278);
             this.button_Cancel.Name = "button_Cancel";
-            this.button_Cancel.Size = new System.Drawing.Size(75, 23);
+            this.button_Cancel.Size = new System.Drawing.Size(75, 24);
             this.button_Cancel.TabIndex = 2;
             this.button_Cancel.Text = "Cancel";
             this.button_Cancel.UseVisualStyleBackColor = true;
@@ -111,18 +116,56 @@ namespace aclogview
             // 
             // groupBox2
             // 
+            this.groupBox2.Controls.Add(this.groupBox4);
             this.groupBox2.Controls.Add(this.displayNodeTooltips);
-            this.groupBox2.Location = new System.Drawing.Point(12, 172);
+            this.groupBox2.Location = new System.Drawing.Point(12, 158);
             this.groupBox2.Name = "groupBox2";
-            this.groupBox2.Size = new System.Drawing.Size(396, 54);
+            this.groupBox2.Size = new System.Drawing.Size(396, 89);
             this.groupBox2.TabIndex = 3;
             this.groupBox2.TabStop = false;
             this.groupBox2.Text = "Treeview";
             // 
+            // groupBox4
+            // 
+            this.groupBox4.Controls.Add(this.label3);
+            this.groupBox4.Controls.Add(this.label4);
+            this.groupBox4.Controls.Add(this.copyAllPadding);
+            this.groupBox4.Location = new System.Drawing.Point(208, 19);
+            this.groupBox4.Name = "groupBox4";
+            this.groupBox4.Size = new System.Drawing.Size(179, 52);
+            this.groupBox4.TabIndex = 4;
+            this.groupBox4.TabStop = false;
+            this.groupBox4.Text = "Pad \"Copy All\" output";
+            // 
+            // label3
+            // 
+            this.label3.AutoSize = true;
+            this.label3.Location = new System.Drawing.Point(11, 25);
+            this.label3.Name = "label3";
+            this.label3.Size = new System.Drawing.Size(48, 13);
+            this.label3.TabIndex = 2;
+            this.label3.Text = "Pad with";
+            // 
+            // label4
+            // 
+            this.label4.AutoSize = true;
+            this.label4.Location = new System.Drawing.Point(104, 25);
+            this.label4.Name = "label4";
+            this.label4.Size = new System.Drawing.Size(43, 13);
+            this.label4.TabIndex = 3;
+            this.label4.Text = "Spaces";
+            // 
+            // copyAllPadding
+            // 
+            this.copyAllPadding.Location = new System.Drawing.Point(65, 22);
+            this.copyAllPadding.Name = "copyAllPadding";
+            this.copyAllPadding.Size = new System.Drawing.Size(33, 20);
+            this.copyAllPadding.TabIndex = 1;
+            // 
             // displayNodeTooltips
             // 
             this.displayNodeTooltips.AutoSize = true;
-            this.displayNodeTooltips.Location = new System.Drawing.Point(6, 19);
+            this.displayNodeTooltips.Location = new System.Drawing.Point(6, 37);
             this.displayNodeTooltips.Name = "displayNodeTooltips";
             this.displayNodeTooltips.Size = new System.Drawing.Size(123, 17);
             this.displayNodeTooltips.TabIndex = 0;
@@ -136,7 +179,7 @@ namespace aclogview
             this.groupBox3.Controls.Add(this.comboBox_TimeFormat);
             this.groupBox3.Location = new System.Drawing.Point(12, 76);
             this.groupBox3.Name = "groupBox3";
-            this.groupBox3.Size = new System.Drawing.Size(396, 90);
+            this.groupBox3.Size = new System.Drawing.Size(396, 76);
             this.groupBox3.TabIndex = 4;
             this.groupBox3.TabStop = false;
             this.groupBox3.Text = "Packets ListView";
@@ -144,11 +187,11 @@ namespace aclogview
             // cbACEStyleHeaders
             // 
             this.cbACEStyleHeaders.AutoSize = true;
-            this.cbACEStyleHeaders.Location = new System.Drawing.Point(6, 55);
+            this.cbACEStyleHeaders.Location = new System.Drawing.Point(9, 46);
             this.cbACEStyleHeaders.Name = "cbACEStyleHeaders";
-            this.cbACEStyleHeaders.Size = new System.Drawing.Size(168, 17);
+            this.cbACEStyleHeaders.Size = new System.Drawing.Size(359, 17);
             this.cbACEStyleHeaders.TabIndex = 1;
-            this.cbACEStyleHeaders.Text = "ACE style packet header flags";
+            this.cbACEStyleHeaders.Text = "ACE style packet header flags (requires reload of file or program restart)";
             this.cbACEStyleHeaders.UseVisualStyleBackColor = true;
             this.cbACEStyleHeaders.CheckedChanged += new System.EventHandler(this.cbACEStyleHeaders_CheckedChanged);
             // 
@@ -180,7 +223,7 @@ namespace aclogview
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.CancelButton = this.button_Cancel;
-            this.ClientSize = new System.Drawing.Size(419, 264);
+            this.ClientSize = new System.Drawing.Size(419, 318);
             this.Controls.Add(this.groupBox3);
             this.Controls.Add(this.groupBox2);
             this.Controls.Add(this.button_Cancel);
@@ -196,6 +239,8 @@ namespace aclogview
             this.groupBox1.PerformLayout();
             this.groupBox2.ResumeLayout(false);
             this.groupBox2.PerformLayout();
+            this.groupBox4.ResumeLayout(false);
+            this.groupBox4.PerformLayout();
             this.groupBox3.ResumeLayout(false);
             this.groupBox3.PerformLayout();
             this.ResumeLayout(false);
@@ -216,6 +261,10 @@ namespace aclogview
         private System.Windows.Forms.GroupBox groupBox3;
         private System.Windows.Forms.Label label2;
         private System.Windows.Forms.ComboBox comboBox_TimeFormat;
+        private System.Windows.Forms.GroupBox groupBox4;
+        private System.Windows.Forms.Label label3;
+        private System.Windows.Forms.Label label4;
+        private System.Windows.Forms.TextBox copyAllPadding;
         private System.Windows.Forms.CheckBox cbACEStyleHeaders;
     }
 }